Moves the TDycore data structure definition into a private header file
Following PETSc data structure, the `_p_TDy` definition is moved into
tdycoreprivate.h.
