+++ src/tdysaturation.c
@@ -1,4 +1,4 @@
-#include "tdycore.h"
+#include <tdycoreprivate.h>
 
 void PressureSaturation_Gardner(PetscReal n,PetscReal m,PetscReal alpha,
                                 PetscReal Pc,PetscReal *Se,PetscReal *dSe_dPc) {
